float learning_rate

+++ clddp/args/train.py
@@ -46,7 +46,7 @@ class RetrievalTrainingArguments(
     per_device_train_batch_size: int = 64
     gradient_accumulation_steps: int = 1
     per_device_eval_batch_size: int = 32
-    learning_rate: int = 2e-5
+    learning_rate: float = 2e-5
     num_train_epochs: int = 5
     warmup_ratio: float = 0.1
     warmup_steps: int = 0
